[engine] improve architecture
* in case scene proxy is not existing yet to update its properties, invoke primitive will wait until it's created

+++ PHEngine/PHEngine/src/Core/GraphicsCore/Renderer/PrimitiveSorter.cpp
@@ -10,6 +10,6 @@ namespace Graphics
     void PrimitiveSorter::SortPrimitivesByOrder(std::vector<std::shared_ptr<PrimitiveSceneProxy>> &primitiveProxies)
     {
         std::sort(primitiveProxies.begin(), primitiveProxies.end(), [](const auto &forwardProxyLeft, const auto &forwardProxyRight)
-                  { return static_cast<uint8_t>(forwardProxyLeft->GetPrimitiveSortOrder()) < static_cast<uint8_t>(forwardProxyRight->GetPrimitiveSortOrder()); });
+                  { return forwardProxyLeft->GetPrimitiveSortOrder() < forwardProxyRight->GetPrimitiveSortOrder(); });
     }
 }
